feat(metrics): Add an add_metrics method to simplify pushing up metrics (#21)
Overview:
We want a one step method of pushing metrics up. add_metrics abstracts away the push method and allow the user to pass multiple metrics through a dictionary.

This also changes the package manager to use Poetry and uses Poetry to distribute the package to PyPI.
Tests are now run in docker containers too.